Replace AddRule with a Concat function
AddRule behaved weirdly with respect to comments, and didn't help when
you needed to Insert or Replace instead of Add, or if you needed to
Add with an Index or Handle. So remove it and provide a Concat method
for assembling the Rule field inside a Rule object instead.

(Also, make Concat support additional data types, so callers don't
need to do `fmt.Sprintf("%d", port)` etc themselves.)