Updates to CI workflows
Remove go version matrix on verify; if gofmt changes, we want to
require the latest version.

Remove gci from the linters because golangci assumes if it's there you
want to use it in preference to gofmt and so formatting problems show
up as gci errors rather than gofmt errors which is confusing.