Allow specifying family/table explicitly per-object
Add new Family and Table fields to all objects; they do not need to be
filled in, because they get defaulted by the transaction code.

Allow passing "" for the family/table to knftables.New, and in that
case, require Family and Table to be filled in explicitly in objects
added to a transaction.

The List* and ParseDump APIs do not currently fill in Family/Table,
for backward-compatibility.